requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
redis==5.0.1
rq==1.15.1
pandas==2.1.3
openpyxl==3.1.2
google-api-python-client==2.108.0
google-auth==2.23.4
aiofiles==0.23.2
python-dotenv==1.0.0
pydantic-settings==2.0.3